dev-AverageSquareMetrePriceAgent: Restructured repository
