+++ Agents/AverageSquareMetrePriceAgent/avgsmpriceagent/agent/__init__.py
@@ -1,0 +1,1 @@
+from .avgprice_estimation import *
