+++ Agents/AverageSquareMetrePriceAgent/avgsmpriceagent/datamodel/__init__.py
