+++ Agents/AverageSquareMetrePriceAgent/avgsmpriceagent/kg_operations/javagateway.py
